requirements.txt
ultralytics>=8.0.0
Pillow>=9.0.0
torch>=2.0.0
gensim>=4.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
